requirements.txt
keras~=3.3.3
mysql-connector-python==8.3.0
pandas~=2.2.0
scikit-learn~=1.5.0
tensorflow~=2.16.1
numpy~=1.26.4
matplotlib~=3.9.0
imblearn~=0.0
seaborn~=0.13.2
scikeras~=0.13.0